.ipynb_checkpoints/requirements-checkpoint.txt
# Core Dependencies
pydantic>=2.0.0

# AWS
boto3>=1.26.0
botocore>=1.29.0

# Document Processing
pypdf2
openpyxl
pandas

# Utilities
pyyaml
jinja2
python-dotenv

# Development (optional)
pytest
black
mypy